[FIX] proper calculation of request_line_ids from active_model
the default_get method supposes that 'active_ids' is the id of the purchase request lines. Which is not the case when the wizard is called from the purchase request. This PR fixes it by testing on 'active_model' and making proper calculations.
See: OCA/purchase-workflow@4fe3697#diff-e458197a92be2847dc330973def5bc39