[FIX] purchase_operating_unit: create PO with warehouse without OU
This commit fixes an issue where, if there is only one warehouse visible
to the user and that warehouse has no Operating Unit (so it's available
for all regardless of OU), PO creation fails with an error.

In case of multiple warehouses being available (with and without OUs)
the one with the OU is preferred and used over the others.